fix(frontend): ETH staking balance adjustment
