+++ frontend/app/src/components/dashboard/OverallBalances.vue
@@ -8,7 +8,7 @@ const { currencySymbol, floatingPrecision } = storeToRefs(useGeneralSettingsStor
 const sessionStore = useSessionSettingsStore();
 const { update } = sessionStore;
 const { timeframe } = storeToRefs(sessionStore);
-const { premium } = storeToRefs(usePremiumStore());
+const premium = usePremium();
 const statistics = useStatisticsStore();
 const { fetchNetValue, getNetValue } = statistics;
 const { totalNetWorth } = storeToRefs(statistics);
