+++ frontend/app/src/components/defi/loan/loans/MakerDaoVaultLoan.vue
@@ -6,7 +6,7 @@ const props = defineProps<{
   vault: MakerDAOVaultModel;
 }>();
 
-const { premium } = storeToRefs(usePremiumStore());
+const premium = usePremium();
 const { t } = useI18n();
 
 const totalInterestOwed = computed<BigNumber>(() => {
